backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
replicate>=0.25.0
reportlab>=4.1.0
cairosvg>=2.7.0
Pillow>=10.0.0
pillow-heif>=0.16.0
requests>=2.31.0
